requirements_dev.txt
bumpversion==0.5.3
wheel==0.23.0
watchdog==0.8.3
flake8==2.4.1
coverage==4.0
Sphinx==1.3.1
cryptography==1.0.1
PyYAML==3.11
pytest
numpy==1.10.4
scipy==0.17.0
pandas==0.17.0
scikit-learn==0.17
